Should not allow deleting of a started task - TRUNK-2790
git-svn-id: http://svn.openmrs.org/openmrs/trunk@24595 5bac5841-c719-aa4e-b3fe-cce5062f897a

+++ web/src/main/java/org/openmrs/scheduler/web/controller/SchedulerListController.java
@@ -112,8 +112,12 @@ public class SchedulerListController extends SimpleFormController {
 					}
 					
 					if (action.equals(msa.getMessage("Scheduler.taskList.delete"))) {
-						schedulerService.deleteTask(Integer.valueOf(taskId));
-						success.append(msa.getMessage("Scheduler.taskList.deleted", args));
+						if (!task.getStarted()) {
+							schedulerService.deleteTask(Integer.valueOf(taskId));
+							success.append(msa.getMessage("Scheduler.taskList.deleted", args));
+						} else {
+							error.append(msa.getMessage("Scheduler.taskList.deleteNotAllowed", args));
+						}
 					} else if (action.equals(msa.getMessage("Scheduler.taskList.stop"))) {
 						schedulerService.shutdownTask(task);
 						success.append(msa.getMessage("Scheduler.taskList.stopped", args));
